+++ ai_hydra/nnet/ATH/ATHGearbox.py
@@ -13,7 +13,7 @@ from __future__ import annotations
 from random import Random
 from dataclasses import dataclass, field
 
-from ai_hydra.constants.DReplayMemory import DMemDef, DMemField, ATH_GEARBOX
+from ai_hydra.constants.DReplayMemory import DMemDef, ATH_GEARBOX
 from ai_hydra.constants.DHydra import DHydraLog, DModule
 from ai_hydra.constants.DHydraTui import DField
 from ai_hydra.constants.DEvent import EV_TYPE
@@ -61,6 +61,12 @@ class ATHGearBox:
         # Handling stagnation by downshifting
         self._stagnation_flag = False
         self._stagnation_alert_count = 0
+        # Track critical stagnation events
+        self._crit_stagnation_alert_count = 0
+
+        # Track when the last upshift happened
+        self._cur_epoch = 0
+        self._last_upshift = 0
 
     async def get_gear(self) -> int:
         self._cur_bucket_counts = bucket_counts = (
@@ -101,7 +107,11 @@ class ATHGearBox:
             await self.event.publish(
                 EventMsg(
                     level=DHydraLog.WARNING,
-                    message=f"Stagnation alert({self._stagnation_alert_count}) - Shifting DOWN: {self._cur_gear + self._stagnation_alert_count} > {self._cur_gear}",
+                    message=(
+                        f"Stagnation alert({self._stagnation_alert_count}) - "
+                        f"Shifting DOWN: {self._cur_gear + self._stagnation_alert_count}"
+                        f" > {self._cur_gear} - {self._cur_seq_length}/{self._cur_batch_size}"
+                    ),
                     ev_type=EV_TYPE.SHIFTING,
                     payload={
                         DField.GEAR: self._cur_gear,
@@ -125,6 +135,7 @@ class ATHGearBox:
                 return self._cur_gear
 
             self._cur_gear += 1
+            self._last_upshift = self._cur_epoch
             self._cooldown_count = 0
             self._cur_seq_length, self._cur_batch_size = ATH_GEARBOX[
                 self._cur_gear
@@ -132,7 +143,7 @@ class ATHGearBox:
             await self.event.publish(
                 EventMsg(
                     level=DHydraLog.INFO,
-                    message=f"Shifting UP: {self._cur_gear - 1} > {self._cur_gear}",
+                    message=f"Shifting UP: {self._cur_gear - 1} > {self._cur_gear} - {self._cur_seq_length}/{self._cur_batch_size}",
                     ev_type=EV_TYPE.SHIFTING,
                     payload={
                         DField.GEAR: self._cur_gear,
@@ -159,7 +170,7 @@ class ATHGearBox:
             await self.event.publish(
                 EventMsg(
                     level=DHydraLog.INFO,
-                    message=f"Shifting DOWN: {self._cur_gear}",
+                    message=f"Shifting DOWN: {self._cur_gear} - {self._cur_seq_length}/{self._cur_batch_size}",
                     ev_type=EV_TYPE.SHIFTING,
                     payload={
                         DField.GEAR: self._cur_gear,
@@ -178,28 +189,44 @@ class ATHGearBox:
         return count
 
     async def hard_reset(self):
-        self._stagnation_flag = False
-        self._cooldown_count = 0
-        old_gear = self._cur_gear
-        self._cur_gear = 3  # seq_length/batch_size == 8/32
-        self._cur_seq_length, self._cur_batch_size = ATH_GEARBOX[
-            self._cur_gear
-        ]
-        await self.event.publish(
-            EventMsg(
-                level=DHydraLog.INFO,
-                message=f"Hard Reset - Shifting DOWN: {old_gear} > {self._cur_gear}",
-                ev_type=EV_TYPE.SHIFTING,
-                payload={
-                    DField.GEAR: self._cur_gear,
-                    DField.SEQ_LENGTH: self._cur_seq_length,
-                    DField.BATCH_SIZE: self._cur_batch_size,
-                },
+        self._crit_stagnation_alert_count += 1
+        max_hard_reset_eps = (
+            DMemDef.MAX_HARD_RESET_EPISODES * self._crit_stagnation_alert_count
+        )
+
+        num_ep_since_upshift = self._cur_epoch - self._last_upshift
+        # Don't do a radical downshift, we just upshifted...
+        if num_ep_since_upshift > max_hard_reset_eps:
+            self._stagnation_flag = False
+            self._cooldown_count = 0
+            old_gear = self._cur_gear
+            self._cur_gear = 3  # seq_length/batch_size == 8/32
+            self._cur_seq_length, self._cur_batch_size = ATH_GEARBOX[
+                self._cur_gear
+            ]
+            await self.event.publish(
+                EventMsg(
+                    level=DHydraLog.INFO,
+                    message=f"Critical Stagnation alert: Radical DOWN shift: {old_gear} > {self._cur_gear} - {self._cur_seq_length}/{self._cur_batch_size}",
+                    ev_type=EV_TYPE.SHIFTING,
+                    payload={
+                        DField.GEAR: self._cur_gear,
+                        DField.SEQ_LENGTH: self._cur_seq_length,
+                        DField.BATCH_SIZE: self._cur_batch_size,
+                    },
+                )
             )
+            return
+
+        self.log.debug(
+            "Received critical stagnation warning: Ignoring because of recent upshift"
         )
 
     def incr_cooldown_count(self):
         self._cooldown_count += 1
+
+    def incr_epoch(self):
+        self._cur_epoch += 1
 
     async def stagnation_cleared(self):
         if self._stagnation_alert_count != 0:
@@ -213,5 +240,16 @@ class ATHGearBox:
         self._stagnation_alert_count = 0
 
     def stagnation_warning(self):
-        self._stagnation_flag = True
-        self._stagnation_alert_count += 1
+        # Don't downshift if we just upshifted....
+        num_ep_since_upshift = self._cur_epoch - self._last_upshift
+        if num_ep_since_upshift > DMemDef.MAX_STAGNANT_EPISODES:
+            self.log.debug(
+                "Received stagnation warning: Setting stagnation flag"
+            )
+            self._stagnation_flag = True
+            self._stagnation_alert_count += 1
+            return
+
+        self.log.debug(
+            "Received stagnation warning: Ignoring because of recent upshift"
+        )