+++ ai_hydra/nnet/TrainMgr.py
@@ -1,19 +1,20 @@
 from __future__ import annotations
 
 from dataclasses import dataclass
-from typing import Optional
 
-from ai_hydra.server.SnakeMgr import SnakeMgr
-from ai_hydra.constants.DHydra import DModule
+from ai_hydra.constants.DHydra import DModule, DHydraLog
+from ai_hydra.constants.DReplayMemory import DMemDef
 
-from ai_hydra.nnet.Policy.HydraPolicy import HydraPolicy
-from ai_hydra.nnet.LinearTrainer import LinearTrainer
-from ai_hydra.nnet.RNNTrainer import RNNTrainer
-from ai_hydra.nnet.SimpleReplayMemory import SimpleReplayMemory
+from ai_hydra.utils.HydraLog import HydraLog
 from ai_hydra.nnet.ATH.ATHMemory import ATHMemory
-from ai_hydra.nnet.Transition import Transition
+from ai_hydra.nnet.LinearTrainer import LinearTrainer
+from ai_hydra.nnet.Policy.HydraPolicy import HydraPolicy
+from ai_hydra.nnet.RecurrentTrainer import RecurrentTrainer
+from ai_hydra.nnet.SimpleReplayMemory import SimpleReplayMemory
 from ai_hydra.nnet.models.LinearModel import LinearModel
 from ai_hydra.nnet.models.RNNModel import RNNModel
+from ai_hydra.nnet.models.GRUModel import GRUModel
+from ai_hydra.server.SnakeMgr import SnakeMgr
 
 
 @dataclass(frozen=True)
@@ -35,8 +36,8 @@ class TrainRunStats:
     avg_loss: float
 
 
-MAX_STAGNANT_EPISODES = 600
-MAX_HARD_RESET_EPISODES = 1500
+MAX_STAGNANT_EPISODES = DMemDef.MAX_STAGNANT_EPISODES
+MAX_HARD_RESET_EPISODES = DMemDef.MAX_HARD_RESET_EPISODES
 
 
 class TrainMgr:
@@ -61,11 +62,18 @@ class TrainMgr:
         *,
         snake_mgr: SnakeMgr,
         policy: HydraPolicy,
-        trainer: LinearTrainer | RNNTrainer,
+        trainer: LinearTrainer | RecurrentTrainer,
         replay: SimpleReplayMemory | ATHMemory,
         client_id: str = DModule.TRAIN_MGR,
-        model: LinearModel | RNNModel,
+        model: LinearModel | RNNModel | GRUModel,
+        log_level: DHydraLog,
     ) -> None:
+
+        self.log = HydraLog(
+            client_id=DModule.TRAIN_MGR,
+            log_level=log_level,
+            to_console=True,
+        )
         self.snake_mgr = snake_mgr
         self.policy = policy
         self.trainer = trainer
@@ -75,9 +83,14 @@ class TrainMgr:
 
         self._stag_ep_count = 0
         self._hard_reset_ep_count = 0
+        self._hard_reset_count = 0
         self._cur_highscore = 0
+        self._max_hard_reset_episodes = MAX_HARD_RESET_EPISODES
 
+    # Called at the end of each episode
     async def handle_stagnation(self, final_score):
+        self.replay.gearbox.incr_epoch()
+
         if final_score > self._cur_highscore:
             self._cur_highscore = final_score
             self._stag_ep_count = 0
@@ -89,10 +102,23 @@ class TrainMgr:
             self._hard_reset_ep_count += 1
 
         if self._stag_ep_count >= MAX_STAGNANT_EPISODES:
+            self.log.debug(
+                "Stagnation event detected, notifying the replay memory"
+            )
             self._stag_ep_count = 0
             self.replay.stagnation_warning()
 
-        if self._hard_reset_ep_count >= MAX_HARD_RESET_EPISODES:
+        if self._hard_reset_ep_count >= self._max_hard_reset_episodes:
+            self._hard_reset_count += 1
+            if self._hard_reset_count > 1:
+                self._max_hard_reset_episodes = (
+                    self._max_hard_reset_episodes * self._hard_reset_count
+                )
+                self.log.debug(
+                    f"Critical stagnation alert({self._hard_reset_count}): "
+                    f"Exceeded {self._hard_reset_count} episodes without "
+                    "improvement, notifying replay memory"
+                )
             self._hard_reset_ep_count = 0
             self._stag_ep_count = 0
             await self.replay.hard_reset()