+++ ai_hydra/nnet/models/GRUModel.py
@@ -1,4 +1,4 @@
-# ai_hydra/net/RNNModel.py
+# ai_hydra/net/GRUModel.py
 #
 #    AI Hydra
 #    Author: Nadim-Daniel Ghaznavi
@@ -10,42 +10,41 @@
 import torch
 import torch.nn as nn
 
-from ai_hydra.constants.DNNet import DNetDef, DRNN
-from ai_hydra.constants.DHydra import DModule, DHydraLog
-from ai_hydra.constants.DHydraTui import DField
+from ai_hydra.constants.DNNet import DNetDef, DGRU
+from ai_hydra.constants.DHydra import DModule
+from ai_hydra.constants.DHydra import DHydraLog
 
 from ai_hydra.utils.HydraLog import HydraLog
 
 
-class RNNModel(nn.Module):
+class GRUModel(nn.Module):
     def __init__(self, log_level: DHydraLog, seed: int):
         super().__init__()
         torch.manual_seed(seed)
 
         self.log = HydraLog(
-            client_id=DModule.RNN_MODEL,
+            client_id=DModule.GRU_MODEL,
             log_level=log_level,
             to_console=True,
         )
 
         self._hidden_size = None
-        self._rnn_dropout = None
-        self._rnn_layers = None
+        self._gru_dropout = None
+        self._gru_layers = None
 
     def _init_model(self):
         input_size = DNetDef.INPUT_SIZE
-        output_size = DRNN.OUTPUT_SIZE
+        output_size = DGRU.OUTPUT_SIZE
 
         self.m_in = nn.Sequential(
             nn.Linear(input_size, self._hidden_size),
             nn.ReLU(),
         )
-        self.m_rnn = nn.RNN(
+        self.m_gru = nn.GRU(
             input_size=self._hidden_size,
             hidden_size=self._hidden_size,
-            nonlinearity=DField.TANH,
-            num_layers=self._rnn_layers,
-            dropout=self._rnn_dropout,
+            num_layers=self._gru_layers,
+            dropout=self._gru_dropout if self._gru_layers > 1 else 0.0,
             batch_first=True,
         )
         self.m_out = nn.Linear(self._hidden_size, output_size)
@@ -57,7 +56,7 @@ class RNNModel(nn.Module):
             x = x.unsqueeze(0)  # [1, T, F]
 
         x = self.m_in(x)  # [B, T, H]
-        x, _ = self.m_rnn(x)  # [B, T, H]
+        x, _ = self.m_gru(x)  # [B, T, H]
         x = self.m_out(x)  # [B, T, A]
 
         return x[:, -1, :]  # [B, A]
@@ -69,7 +68,7 @@ class RNNModel(nn.Module):
             x = x.unsqueeze(0)  # [1, T, F]
 
         x = self.m_in(x)  # [B, T, H]
-        x, _ = self.m_rnn(x)  # [B, T, H]
+        x, _ = self.m_gru(x)  # [B, T, H]
         x = self.m_out(x)  # [B, T, A]
         return x
 
@@ -81,12 +80,12 @@ class RNNModel(nn.Module):
         self.apply(_reset)
 
     def set_params(
-        self, hidden_size: int, dropout_p: float, rnn_layers: int
+        self, hidden_size: int, dropout_p: float, gru_layers: int
     ) -> None:
         self._hidden_size = hidden_size
-        self._rnn_dropout = dropout_p
-        self._rnn_layers = rnn_layers
+        self._gru_dropout = dropout_p
+        self._gru_layers = gru_layers
         self.log.info(f"Setting hidden size to {hidden_size}")
         self.log.info(f"Setting dropout layer p-value to {dropout_p}")
-        self.log.info(f"Setting the number of RNN layers to {rnn_layers}")
+        self.log.info(f"Setting the number of GRU layers to {gru_layers}")
         self._init_model()